feat(contracts-communication): use uint64 for chain IDs and DB nonces (#2453)
* Update chainId constants to be uint64

* Update `srcChainId` references

* Update `dstChainId` references

* Update `chainId` references

* Update `remoteChainId` references

* Fix MessageBus w/o breaking interface

* Fix other contracts

* Update nonce constants to be uint64

* Update `dbNonce` to be uint64

* Update tests

* Add InterchainTx Header compression

* Use compact encoding for ic transactions

* Add InterchainBatch Key compression

* Use compact encoding for batches

* Use new BatchKey type for DB mapping

* Cleanup unused/deprecated functions

* Go changes for #2453 (#2454)

* Regenerate `committee`

* Fix committee E2E test

* Regenerate `sin-executor`

* Fix broken casts in `sin-executor`

* Fix sin-executor tests

* Update PingPong integration tests to use uint64

* Revert "Regenerate `sin-executor`"

This reverts commit 01fa75c9b8dacfd695c920515c8ff631f3770b36.

* Revert "Regenerate `committee`"

This reverts commit e096bb2e07df72f409045d9b7c75bdfc0c349e7f.

* Fix: forgotten uint256 indexed chainId fields

* Regenerate `committee`

* Update `committee` with uint256 -> uint64 change

* Regenerate `sin-executor`

+++ packages/contracts-communication/contracts/InterchainClientV1.sol
@@ -39,7 +39,7 @@ contract InterchainClientV1 is Ownable, InterchainClientV1Events, IInterchainCli
     address public executionFees;
 
     /// @dev Address of the InterchainClient contract on the remote chain
-    mapping(uint256 chainId => bytes32 remoteClient) internal _linkedClient;
+    mapping(uint64 chainId => bytes32 remoteClient) internal _linkedClient;
     /// @dev Executor address that completed the transaction. Address(0) if not executed yet.
     mapping(bytes32 transactionId => address executor) internal _txExecutor;
 
@@ -54,14 +54,14 @@ contract InterchainClientV1 is Ownable, InterchainClientV1Events, IInterchainCli
     }
 
     // @inheritdoc IInterchainClientV1
-    function setLinkedClient(uint256 chainId, bytes32 client) external onlyOwner {
+    function setLinkedClient(uint64 chainId, bytes32 client) external onlyOwner {
         _linkedClient[chainId] = client;
         emit LinkedClientSet(chainId, client);
     }
 
     // @inheritdoc IInterchainClientV1
     function interchainSend(
-        uint256 dstChainId,
+        uint64 dstChainId,
         bytes32 receiver,
         address srcExecutionService,
         address[] calldata srcModules,
@@ -77,7 +77,7 @@ contract InterchainClientV1 is Ownable, InterchainClientV1Events, IInterchainCli
 
     // @inheritdoc IInterchainClientV1
     function interchainSendEVM(
-        uint256 dstChainId,
+        uint64 dstChainId,
         address receiver,
         address srcExecutionService,
         address[] calldata srcModules,
@@ -132,7 +132,7 @@ contract InterchainClientV1 is Ownable, InterchainClientV1Events, IInterchainCli
     }
 
     /// @inheritdoc IInterchainClientV1
-    function writeExecutionProof(bytes32 transactionId) external returns (uint256 dbNonce, uint64 entryIndex) {
+    function writeExecutionProof(bytes32 transactionId) external returns (uint64 dbNonce, uint64 entryIndex) {
         address executor = _txExecutor[transactionId];
         if (executor == address(0)) {
             revert InterchainClientV1__TxNotExecuted(transactionId);
@@ -166,7 +166,7 @@ contract InterchainClientV1 is Ownable, InterchainClientV1Events, IInterchainCli
 
     // @inheritdoc IInterchainClientV1
     function getInterchainFee(
-        uint256 dstChainId,
+        uint64 dstChainId,
         address srcExecutionService,
         address[] calldata srcModules,
         bytes calldata options,
@@ -189,7 +189,7 @@ contract InterchainClientV1 is Ownable, InterchainClientV1Events, IInterchainCli
     }
 
     /// @inheritdoc IInterchainClientV1
-    function getLinkedClient(uint256 chainId) external view returns (bytes32) {
+    function getLinkedClient(uint64 chainId) external view returns (bytes32) {
         if (chainId == block.chainid) {
             revert InterchainClientV1__NotRemoteChainId(chainId);
         }
@@ -197,7 +197,7 @@ contract InterchainClientV1 is Ownable, InterchainClientV1Events, IInterchainCli
     }
 
     /// @inheritdoc IInterchainClientV1
-    function getLinkedClientEVM(uint256 chainId) external view returns (address linkedClientEVM) {
+    function getLinkedClientEVM(uint64 chainId) external view returns (address linkedClientEVM) {
         if (chainId == block.chainid) {
             revert InterchainClientV1__NotRemoteChainId(chainId);
         }
@@ -226,7 +226,7 @@ contract InterchainClientV1 is Ownable, InterchainClientV1Events, IInterchainCli
 
     /// @dev Internal logic for sending a message to another chain.
     function _interchainSend(
-        uint256 dstChainId,
+        uint64 dstChainId,
         bytes32 receiver,
         address srcExecutionService,
         address[] calldata srcModules,
@@ -257,7 +257,7 @@ contract InterchainClientV1 is Ownable, InterchainClientV1Events, IInterchainCli
         desc.transactionId = keccak256(encodeTransaction(icTx));
         // Sanity check: nonce returned from DB should match the nonce used to construct the transaction
         {
-            (uint256 dbNonce, uint64 entryIndex) = IInterchainDB(INTERCHAIN_DB).writeEntryWithVerification{
+            (uint64 dbNonce, uint64 entryIndex) = IInterchainDB(INTERCHAIN_DB).writeEntryWithVerification{
                 value: verificationFee
             }(icTx.dstChainId, desc.transactionId, srcModules);
             assert(dbNonce == desc.dbNonce && entryIndex == desc.entryIndex);
@@ -334,7 +334,7 @@ contract InterchainClientV1 is Ownable, InterchainClientV1Events, IInterchainCli
     }
 
     /// @dev Asserts that the chain is linked and returns the linked client address.
-    function _assertLinkedClient(uint256 chainId) internal view returns (bytes32 linkedClient) {
+    function _assertLinkedClient(uint64 chainId) internal view returns (bytes32 linkedClient) {
         if (chainId == block.chainid) {
             revert InterchainClientV1__NotRemoteChainId(chainId);
         }
